refactor: extract claimTask logic to service layer
