+++ src/tools/claimTaskTool.ts
@@ -1,11 +1,5 @@
-import { isClaimable } from "../models";
-import { TrellisObject } from "../models/TrellisObject";
-import { TrellisObjectStatus } from "../models/TrellisObjectStatus";
-import { TrellisObjectType } from "../models/TrellisObjectType";
 import { Repository } from "../repositories";
-import { checkPrerequisitesComplete } from "../utils/checkPrerequisitesComplete";
-import { filterUnavailableObjects } from "../utils/filterUnavailableObjects";
-import { sortTrellisObjects } from "../utils/sortTrellisObjects";
+import { TaskTrellisService } from "../services/TaskTrellisService";
 
 export const claimTaskTool = {
   name: "claim_task",
@@ -59,7 +53,11 @@ Essential for autonomous task execution workflows where agents need to discover 
   },
 } as const;
 
-export async function handleClaimTask(repository: Repository, args: unknown) {
+export async function handleClaimTask(
+  service: TaskTrellisService,
+  repository: Repository,
+  args: unknown,
+) {
   const {
     scope,
     taskId,
@@ -70,164 +68,5 @@ export async function handleClaimTask(repository: Repository, args: unknown) {
     force?: boolean;
   };
 
-  try {
-    let claimedTask: TrellisObject;
-
-    if (taskId) {
-      claimedTask = await claimSpecificTask(taskId, force, repository);
-    } else {
-      claimedTask = await findNextAvailableTask(scope, repository);
-    }
-
-    const updatedTask = await updateTaskStatus(claimedTask, repository);
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: `Successfully claimed task: ${JSON.stringify(updatedTask, null, 2)}`,
-        },
-      ],
-    };
-  } catch (error) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: `Error claiming task: ${error instanceof Error ? error.message : String(error)}`,
-        },
-      ],
-    };
-  }
-}
-
-async function validateTaskForClaiming(
-  task: TrellisObject,
-  taskId: string,
-  force: boolean,
-  repository: Repository,
-): Promise<void> {
-  // Validate it's a task
-  if (task.type !== TrellisObjectType.TASK) {
-    throw new Error(
-      `Object with ID "${taskId}" is not a task (type: ${task.type})`,
-    );
-  }
-
-  if (!force) {
-    // Validate status is draft or open
-    if (!isClaimable(task)) {
-      throw new Error(
-        `Task "${taskId}" cannot be claimed (status: ${task.status}). Task must be in draft or open status.`,
-      );
-    }
-
-    // Validate all prerequisites are complete
-    const prerequisitesComplete = await checkPrerequisitesComplete(
-      task,
-      repository,
-    );
-    if (!prerequisitesComplete) {
-      throw new Error(
-        `Task "${taskId}" cannot be claimed. Not all prerequisites are complete.`,
-      );
-    }
-  }
-}
-
-async function claimSpecificTask(
-  taskId: string,
-  force: boolean,
-  repository: Repository,
-): Promise<TrellisObject> {
-  const object = await repository.getObjectById(taskId);
-
-  if (!object) {
-    throw new Error(`Task with ID "${taskId}" not found`);
-  }
-
-  await validateTaskForClaiming(object, taskId, force, repository);
-  return object;
-}
-
-async function findNextAvailableTask(
-  scope: string | undefined,
-  repository: Repository,
-): Promise<TrellisObject> {
-  const objects = await repository.getObjects(
-    false, // includeClosed
-    scope,
-    TrellisObjectType.TASK,
-  );
-
-  // Filter to get only available tasks
-  const availableTasks = filterUnavailableObjects(objects);
-
-  if (availableTasks.length === 0) {
-    throw new Error("No available tasks to claim");
-  }
-
-  // Sort by priority and return the top one
-  const sortedTasks = sortTrellisObjects(availableTasks);
-  return sortedTasks[0];
-}
-
-async function updateTaskStatus(
-  task: TrellisObject,
-  repository: Repository,
-): Promise<TrellisObject> {
-  const updatedTask = {
-    ...task,
-    status: TrellisObjectStatus.IN_PROGRESS,
-  };
-
-  await repository.saveObject(updatedTask);
-
-  // Update parent hierarchy to in-progress (don't let errors fail the task claim)
-  try {
-    await updateParentHierarchy(task.parent, repository);
-  } catch (error) {
-    // Log but don't propagate parent hierarchy update errors
-    console.warn("Failed to update parent hierarchy:", error);
-  }
-
-  return updatedTask;
-}
-
-async function updateParentHierarchy(
-  parentId: string | undefined,
-  repository: Repository,
-  visitedIds: Set<string> = new Set(),
-): Promise<void> {
-  if (!parentId) {
-    return;
-  }
-
-  // Prevent infinite recursion by checking if we've already visited this ID
-  if (visitedIds.has(parentId)) {
-    return;
-  }
-  visitedIds.add(parentId);
-
-  const parent = await repository.getObjectById(parentId);
-  if (!parent) {
-    return;
-  }
-
-  // If parent is already in progress, we can stop here since we assume
-  // its parent is already in progress too
-  if (parent.status === TrellisObjectStatus.IN_PROGRESS) {
-    return;
-  }
-
-  // Update parent to in-progress
-  const updatedParent = {
-    ...parent,
-    status: TrellisObjectStatus.IN_PROGRESS,
-  };
-
-  await repository.saveObject(updatedParent);
-
-  // Continue up the hierarchy
-  await updateParentHierarchy(parent.parent, repository, visitedIds);
+  return service.claimTask(repository, scope, taskId, force);
 }
